requirements.txt
ultralytics>=8.0.0
gtts>=2.3.0
pygame>=2.0.0
opencv-python>=4.5.0
pillow>=9.0.0
moviepy>=1.0.0
easyocr>=1.6.0
segmentation-models-pytorch>=0.3.0
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0